# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Tournament API
  description: API for managing tournaments with hexagonal architecture
  version: 1.0.0
  contact:
    name: Tournament API Team

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Tournaments
    description: Tournament management endpoints
  - name: Teams
    description: Team management endpoints

paths:
  /api/tournaments:
    get:
      tags:
        - Tournaments
      summary: Get all tournaments
      description: Retrieve a list of all tournaments
      operationId: getAllTournaments
      responses:
        '200':
          description: List of tournaments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TournamentResponse'

    post:
      tags:
        - Tournaments
      summary: Create a new tournament
      description: Create a new tournament with name and description
      operationId: createTournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentRequest'
      responses:
        '201':
          description: Tournament created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
        '400':
          description: Invalid request body or duplicate tournament name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tournaments/{tournamentId}:
    get:
      tags:
        - Tournaments
      summary: Get tournament by ID
      description: Retrieve a specific tournament by its ID
      operationId: getTournamentById
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: Tournament retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Tournaments
      summary: Update tournament
      description: Update an existing tournament's name and description
      operationId: updateTournament
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TournamentRequest'
      responses:
        '200':
          description: Tournament updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
        '400':
          description: Invalid request body or duplicate tournament name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tournaments
      summary: Delete tournament
      description: Delete a tournament (only allowed if status is CREATED or CANCELLED)
      operationId: deleteTournament
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '204':
          description: Tournament deleted successfully
        '400':
          description: Tournament cannot be deleted (invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tournaments/{tournamentId}/start:
    patch:
      tags:
        - Tournaments
      summary: Start tournament
      description: Change tournament status from CREATED to IN_PROGRESS
      operationId: startTournament
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: Tournament started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
        '400':
          description: Tournament cannot be started (invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tournaments/{tournamentId}/end:
    patch:
      tags:
        - Tournaments
      summary: End tournament
      description: Change tournament status from IN_PROGRESS to COMPLETED
      operationId: endTournament
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: Tournament ended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
        '400':
          description: Tournament cannot be ended (invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/tournaments/{tournamentId}/cancel:
    patch:
      tags:
        - Tournaments
      summary: Cancel tournament
      description: Change tournament status to CANCELLED (only allowed if not COMPLETED)
      operationId: cancelTournament
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: Tournament cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentResponse'
        '400':
          description: Tournament cannot be cancelled (already completed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /{tournamentId}/teams:
    get:
      tags:
        - Teams
      summary: Get all teams in a tournament
      description: Retrieve a list of all teams participating in a specific tournament
      operationId: getAllTeams
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      responses:
        '200':
          description: List of teams retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Teams
      summary: Create a new team in a tournament
      description: Create a new team with name and coach for a specific tournament
      operationId: createTeam
      parameters:
        - $ref: '#/components/parameters/TournamentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
      responses:
        '201':
          description: Team created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request body or duplicate team name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /{tournamentId}/teams/{teamId}:
    get:
      tags:
        - Teams
      summary: Get team by ID
      description: Retrieve a specific team by its ID in a tournament
      operationId: getTeamById
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '404':
          description: Tournament or team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Teams
      summary: Update team
      description: Update an existing team's name and coach in a tournament
      operationId: updateTeam
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
      responses:
        '200':
          description: Team updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: Invalid request body or duplicate team name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Tournament or team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Teams
      summary: Delete team
      description: Delete a team from a tournament
      operationId: deleteTeam
      parameters:
        - $ref: '#/components/parameters/TournamentId'
        - $ref: '#/components/parameters/TeamId'
      responses:
        '204':
          description: Team deleted successfully
        '404':
          description: Tournament or team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    TournamentId:
      name: tournamentId
      in: path
      description: Tournament unique identifier
      required: true
      schema:
        type: integer
        format: int64
        example: 1

    TeamId:
      name: teamId
      in: path
      description: Team unique identifier
      required: true
      schema:
        type: integer
        format: int64
        example: 1

  schemas:
    TournamentRequest:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Tournament name (must be unique)
          example: Copa America 2025
        description:
          type: string
          minLength: 10
          maxLength: 500
          description: Tournament description
          example: Torneo internacional de futbol en Sudamerica

    TournamentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Tournament unique identifier
          example: 1
        name:
          type: string
          description: Tournament name
          example: Copa America 2025
        description:
          type: string
          description: Tournament description
          example: Torneo internacional de futbol en Sudamerica
        status:
          $ref: '#/components/schemas/TournamentStatus'

    TournamentStatus:
      type: string
      enum:
        - CREATED
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED
      description: |
        Tournament status lifecycle:
        - CREATED: Initial state when tournament is created
        - IN_PROGRESS: Tournament has started
        - COMPLETED: Tournament has finished
        - CANCELLED: Tournament was cancelled (can be cancelled from any state except COMPLETED)

    TeamRequest:
      type: object
      required:
        - name
        - coach
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Team name (must be unique within tournament)
          example: Real Madrid
        coach:
          type: string
          minLength: 3
          maxLength: 100
          description: Team coach name
          example: Carlo Ancelotti

    TeamResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Team unique identifier
          example: 1
        name:
          type: string
          description: Team name
          example: Real Madrid
        coach:
          type: string
          description: Team coach name
          example: Carlo Ancelotti
        tournamentId:
          type: integer
          format: int64
          description: Tournament identifier
          example: 1
        points:
          type: integer
          description: Total points earned
          example: 9
        matchesPlayed:
          type: integer
          description: Number of matches played
          example: 3
        matchesWin:
          type: integer
          description: Number of victories
          example: 3
        matchesDraw:
          type: integer
          description: Number of draws
          example: 0
        matchesLost:
          type: integer
          description: Number of defeats
          example: 0
        goalsFor:
          type: integer
          description: Goals scored by the team
          example: 8
        goalsAgainst:
          type: integer
          description: Goals conceded by the team
          example: 2
        goalDifference:
          type: integer
          description: Goal difference (goalsFor - goalsAgainst)
          example: 6

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Tournament not found
        path:
          type: string
          description: Request path
          example: /api/tournaments/1